+++ src/plonemeeting/portal/core/content/publication.py
@@ -1,11 +1,7 @@
 from AccessControl import ClassSecurityInfo
-from DateTime import DateTime
-from Products.CMFCore.permissions import ManagePortal
-from Products.CMFCore.permissions import ModifyPortalContent
-from Products.CMFCore.permissions import View
-from Products.CMFCore.utils import _checkPermission
 from collective.timestamp.behaviors.timestamp import ITimestampableDocument
 from collective.timestamp.interfaces import ITimeStamper
+from DateTime import DateTime
 from imio.helpers import EMPTY_DATETIME
 from imio.helpers.content import object_values
 from plone import api
@@ -23,6 +19,10 @@ from plone.indexer.decorator import indexer
 from plone.namedfile.field import NamedBlobFile
 from plone.supermodel import model
 from plonemeeting.portal.core import _
+from Products.CMFCore.permissions import ManagePortal
+from Products.CMFCore.permissions import ModifyPortalContent
+from Products.CMFCore.permissions import View
+from Products.CMFCore.utils import _checkPermission
 from zope import schema
 from zope.component import getMultiAdapter
 from zope.globalrequest import getRequest
@@ -159,23 +159,8 @@ class Publication(Container, File):
         )
 
     def may_publish(self):
-        """May publish if able to modify and
-        a "publication date" (effectiveDate) is NOT defined.
-        When "unpublished" check that time stamp  was not modified."""
-        return _checkPermission(ModifyPortalContent, self) and (
-            self.effective_date is None
-            or (api.content.get_state(self) == "planned" and self.effective_date < DateTime())
-            or (api.content.get_state(self) == "unpublished" and self.is_power_user() and self.timestamp_still_valid())
-            or (
-                api.content.get_state(self) == "private"
-                and self.effective_date < DateTime()
-                and self.enable_timestamping is False
-            )
-        )
-
-    def timestamp_still_valid(self):
-        """Check if timestamp still corresponds to effective_date."""
-        return True
+        """May publish if able to modify."""
+        return _checkPermission(ModifyPortalContent, self)
 
 
 @indexer(IPublication)