+++ src/plonemeeting/portal/core/tests/test_item.py
@@ -24,7 +24,7 @@ class TestItemView(PmPortalDemoFunctionalTestCase):
         # project disclaimer message not displayed
         self.assertFalse("alert-content" in view())
         # when meeting is in decision
-        self.login_as_manager()
+        self.login_as_admin()
         api.content.transition(self.meeting, to_state="in_project")
         self.assertEqual(api.content.get_state(self.meeting), "in_project")
         view = self.item.restrictedTraverse("@@view")
@@ -45,7 +45,7 @@ class TestItemView(PmPortalDemoFunctionalTestCase):
         self.assertSetEqual({"previous_item", "next_item"}, set(next_prev_infos.keys()))
 
         self.assertSetEqual(
-            {"id", "title", "description", "portal_type", "url"}, set(next_prev_infos["next_item"].keys())
+            {"id", "title", "description", "portal_type", "url", "obj"}, set(next_prev_infos["next_item"].keys())
         )
 
         # Moving item should not change the last item number and the next/previous items